dom/media/tests/mochitest/test_peerConnection_bug827843.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="pc.js"></script>
  <script type="application/javascript" src="templates.js"></script>
  <script type="application/javascript" src="turnConfig.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    bug: "827843",
    title: "Ensure that localDescription and remoteDescription are null after close"
  });

  var steps = [
    [
      "CHECK_SDP_ON_CLOSED_PC",
      function (test) {
        var description;
        var exception = null;

        // handle the event which the close() triggers
        test.pcLocal.onsignalingstatechange = function (state) {
          is(state, "closed", "Received expected onsignalingstatechange event 'closed'");
        }

        test.pcLocal.close();

        try { description = test.pcLocal.localDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access localDescription of pcLocal after close throws exception");
        exception = null;

        try { description = test.pcLocal.remoteDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access remoteDescription of pcLocal after close throws exception");
        exception = null;

        // handle the event which the close() triggers
        test.pcRemote.onsignalingstatechange = function (state) {
          is(state, "closed", "Received expected onsignalingstatechange event 'closed'");
        }

        test.pcRemote.close();

        try  { description = test.pcRemote.localDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access localDescription of pcRemote after close throws exception");
        exception = null;

        try  { description = test.pcRemote.remoteDescription; } catch (e) { exception = e; }
        ok(exception, "Attempt to access remoteDescription of pcRemote after close throws exception");

        test.next();
      }
    ]
  ];

  var test;
  runNetworkTest(function () {
    test = new PeerConnectionTest();
    test.setMediaConstraints([{audio: true}], [{audio: true}]);
    test.chain.append(steps);
    test.run();
  });
</script>
</pre>
</body>
</html>
